Make 0.9.7 compilable on Windows and for 32 bit x86 targets

+++ src/rpcblockchain.cpp
@@ -356,7 +356,7 @@ Object blockToJSON(const CBlock& block, const CBlockIndex* blockindex)
       
     }
     result.push_back(Pair("SSF height",get_ssf_height(blockindex)));
-    result.push_back(Pair("SSF work",get_ssf_work(blockindex)));
+    result.push_back(Pair("SSF work", (int64_t)get_ssf_work(blockindex)));
     result.push_back(Pair("SSF time",get_ssf_time(blockindex)));
     result.push_back(Pair("merkleroot", block.hashMerkleRoot.GetHex()));
     Array txs;
